backend/requirements.txt
Django==4.1.13
djongo==1.3.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
stripe==7.6.0
pymongo==3.12.3
pytz
python-dotenv
requests==2.31.0
redis==5.0.1
django-redis==5.4.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
gunicorn==21.2.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
